+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - context: add `Context::try_send`.
 - errors: add `TrySendError::map` and `RequestError::map`.
 - dumping: write `ts` firstly to support the `sort` utility.
+- dumper: extract a dump's name if it's not specified.
 
 ### Fixed
 - init: do not start termination if the memory tracker fails to read files.